Rename and move configuration module.
