+++ src/ImageSharp/Formats/Bmp/BmpConfigurationModule.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors.
+// Copyright (c) Six Labors.
 // Licensed under the Six Labors Split License.
 
 namespace SixLabors.ImageSharp.Formats.Bmp;
@@ -6,7 +6,7 @@ namespace SixLabors.ImageSharp.Formats.Bmp;
 /// <summary>
 /// Registers the image encoders, decoders and mime type detectors for the bmp format.
 /// </summary>
-public sealed class BmpConfigurationModule : IConfigurationModule
+public sealed class BmpConfigurationModule : IImageFormatConfigurationModule
 {
     /// <inheritdoc/>
     public void Configure(Configuration configuration)
