+++ src/ImageSharp/Formats/Gif/GifConfigurationModule.cs
@@ -6,7 +6,7 @@ namespace SixLabors.ImageSharp.Formats.Gif;
 /// <summary>
 /// Registers the image encoders, decoders and mime type detectors for the gif format.
 /// </summary>
-public sealed class GifConfigurationModule : IConfigurationModule
+public sealed class GifConfigurationModule : IImageFormatConfigurationModule
 {
     /// <inheritdoc/>
     public void Configure(Configuration configuration)
