+++ src/ImageSharp/Formats/Tiff/TiffConfigurationModule.cs
@@ -6,7 +6,7 @@ namespace SixLabors.ImageSharp.Formats.Tiff;
 /// <summary>
 /// Registers the image encoders, decoders and mime type detectors for the TIFF format.
 /// </summary>
-public sealed class TiffConfigurationModule : IConfigurationModule
+public sealed class TiffConfigurationModule : IImageFormatConfigurationModule
 {
     /// <inheritdoc/>
     public void Configure(Configuration configuration)
