+++ tests/ImageSharp.Tests/ConfigurationTests.cs
@@ -3,6 +3,7 @@
 
 using Microsoft.DotNet.RemoteExecutor;
 using Moq;
+using SixLabors.ImageSharp.Formats;
 using SixLabors.ImageSharp.Formats.Bmp;
 using SixLabors.ImageSharp.IO;
 using SixLabors.ImageSharp.Memory;
@@ -86,7 +87,7 @@ public class ConfigurationTests
     [Fact]
     public void ConstructorCallConfigureOnFormatProvider()
     {
-        var provider = new Mock<IConfigurationModule>();
+        var provider = new Mock<IImageFormatConfigurationModule>();
         var config = new Configuration(provider.Object);
 
         provider.Verify(x => x.Configure(config));
@@ -95,7 +96,7 @@ public class ConfigurationTests
     [Fact]
     public void AddFormatCallsConfig()
     {
-        var provider = new Mock<IConfigurationModule>();
+        var provider = new Mock<IImageFormatConfigurationModule>();
         var config = new Configuration();
         config.Configure(provider.Object);
 
@@ -178,7 +179,7 @@ public class ConfigurationTests
         }
     }
 
-    private class MockConfigurationModule : IConfigurationModule
+    private class MockConfigurationModule : IImageFormatConfigurationModule
     {
         public void Configure(Configuration configuration)
         {
